Remove assigning unnecessary variables in api middleware
The api results are copied two times to rebuild them for internal use. This is
unnecessary and should be prevented by using the Array.map() method.

The logic was merely moved and the lines of code that were commented out were
left untouched. Normally though committing lines of code that are commented out
is bad form and should be avoided.