Prevent fetching all types from the API instead of just the one
The "api" middleware is just used to fill the whole store (all types) when it is
loaded. That is unnecessary because just the data from one type is used.
The "api" middleware should be changed to just add a function to the context
that can be used to fill the store for one type from the API.

Because the endpoints and return value of all three types is so similar this can
be done with one generic "fetch" function.